Solve occasional segfault
The cause of the segfault is unknown.
It occurred on some values of the size parameter,
even when that parameter was not actually used in the function,
and whether it occurred
also depended on the presence of the docstring,
whether the size parameter was passed literally or as a computation,
and whether there was a sleep or print operation nearby the call.

The segfaulting access was determined to be on PRNG state, however.
Allocating the state on the stack solved the issue,
and has minimal performance impact.